Fix bug with extension loader when DISCARD ALL is executed
Previously, the GUC to let the versioned-extension know that
the loader was present was incorrectly reset during DISCARD ALL.
This caused newly minted parallel workers to throw an error
about the extension not being preloaded since it did not
know that a loader was, in fact, preloaded. This PR
fixes this issue by changing the GUC to a rendezvous variable.

It also disables the loader in parallel workers since the loading
should have been handled in the parallel leader and the worker
start-up logic.